investigations/requirements.txt
# Event Poster Analyzer Dependencies
# Core analysis and AI
anthropic>=0.34.0           # Claude API client
pydantic>=2.5.0             # Data validation and serialization

# Image processing and QR detection
Pillow>=10.0.0              # Image processing
opencv-python>=4.8.0        # Computer vision for image preprocessing
pyzbar>=0.1.9               # QR code detection
numpy>=1.24.0               # Numerical operations for image processing

# Web scraping and HTTP requests  
aiohttp>=3.9.0              # Async HTTP client for web scraping
beautifulsoup4>=4.12.0      # HTML parsing
requests>=2.31.0            # HTTP requests (fallback)
validators>=0.22.0          # URL and data validation

# German-specific processing
python-dateutil>=2.8.0     # Date parsing with German format support  
phonenumbers>=8.13.0       # German phone number validation

# Testing framework
pytest>=7.4.0              # Testing framework
pytest-asyncio>=0.21.0     # Async test support
pytest-mock>=3.12.0        # Mocking utilities

# Optional: Enhanced OCR (uncomment if using pytesseract)
# pytesseract>=0.3.10
# tesseract>=0.1.3

# Development and code quality (optional)
# black>=23.9.0             # Code formatting
# isort>=5.12.0             # Import sorting  
# flake8>=6.1.0             # Code linting
# mypy>=1.6.0               # Type checking